No more libswd
That was a silly library, full of middleware and memory leaks. Debugging libswd was a good way to get familiar enough with SWD to write a replacement really quickly :P

This still doesn't work reliably, it seems to have the same unsynchronization bug that the old code was having. But it's much smaller and easier to understand now!