Actually, disable linearCutoff by default
When staring directly at the LEDs, the dithering flicker is visible- but in an enclosure, the flicker is totally nonexistent and the linear section causes a noticeable discontinuity.

A pure nonlinear-only approach seems like the best default. I'll document cases when this linear section might help, and it will be up to individual clients (or fcserver config files) to enable it.